Allow an event to be saved and enqueued in the same batch.
putting it on the queue within the same batch.
